pyproject.toml
[project]
name = "ai-admin"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "kerb>=0.2.1",
    "anthropic>=0.40.0",
    "python-telegram-bot[job-queue]>=21.0",
    "mcp",
    "asyncpg",
    "httpx",
    "python-dotenv",
    "pydantic>=2.0",
    "structlog",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot", "agent", "tools"]

[tool.uv]
dev-dependencies = []
